New script from cubiq.org
It is working very well! But I had to test it on my iPad or iPhone for
it to work as it is not looking for the USERAGENT string. So test it on
a live site when trying as it did not work via. localhost and a user
agent spoofer.

Added a few more settings as well that we needed.

Still miss to add something to the metafield in header and pass
settings into js. Also we need the less file to work ;)